fix: rename provider
